Added i64, and changed default int to i32

+++ Valestrom/Astronomer/src/net/verdagon/vale/astronomer/rules.scala
@@ -118,7 +118,7 @@ case class IsaAR(
 sealed trait ITemplexA {
   def resultType: ITemplataType
 }
-case class IntAT(range: RangeS, value: Int) extends ITemplexA {
+case class IntAT(range: RangeS, value: Long) extends ITemplexA {
   override def resultType: ITemplataType = IntegerTemplataType
 }
 case class StringAT(range: RangeS, value: String) extends ITemplexA {
